много наворочено, возможны сбои настроек и баги миграций. Панель карты.карьеры политики звуки офис-менеджер кулер поломки

+++ Assets/Scripts/UI/TeamMemberCardUI.cs
@@ -92,7 +92,7 @@ public class TeamMemberCardUI : MonoBehaviour
                 int xpForCurrentRank = assignedStaff.currentRank.experienceRequired;
                 int xpForNextRank = nextRankData.experienceRequired;
                 int totalXpForLevel = xpForNextRank - xpForCurrentRank;
-                int currentXpInLevel = assignedStaff.experiencePoints - xpForCurrentRank;
+                int currentXpInLevel = (int)assignedStaff.experiencePoints - xpForCurrentRank;
                 xpBarFill.fillAmount = totalXpForLevel > 0 ? (float)currentXpInLevel / totalXpForLevel : 1f;
                 xpText.text = $"XP: {currentXpInLevel} / {totalXpForLevel}";
             }
